chapter 6
I really flubbed this one. you can see compared to the instructor's files (one is txt to avoid overwrite) that I went in and started changing a bunch of code because I thought that the add_student/add_course/add_grade functions were the places to add my input stream. I was wrong. the instructor wrote a whole new initialize function (in the codedemo.cpp) and used the add_ functions to write them in. I really messed this up and came at it from the completely wrong way.